fix: refactorings and nitpicks
