+++ src/main/java/dev/amble/ait/compat/jei/AITJeiPlugin.java
@@ -1,6 +1,7 @@
 package dev.amble.ait.compat.jei;
 
 import dev.amble.ait.AITMod;
+import dev.amble.ait.compat.jei.client.FabricatorRecipeCategory;
 import dev.amble.ait.core.AITBlocks;
 import dev.amble.ait.core.AITItems;
 import dev.amble.ait.core.item.blueprint.BlueprintRegistry;
@@ -8,12 +9,13 @@ import mezz.jei.api.IModPlugin;
 import mezz.jei.api.JeiPlugin;
 import mezz.jei.api.registration.*;
 import net.minecraft.util.Identifier;
+import org.jetbrains.annotations.NotNull;
 
 @JeiPlugin
 public class AITJeiPlugin implements IModPlugin {
 
     @Override
-    public Identifier getPluginUid() {
+    public @NotNull Identifier getPluginUid() {
         return new Identifier(AITMod.MOD_ID, "jei_plugin");
     }
